# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Good Deal - Telecom Bundle Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  phone: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  phone: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Operator {
  id: number;
  name: string;
  slug: string;
  color: string;
  /** @nullable */
  logoUrl: string | null;
  active: boolean;
}

export interface Bundle {
  id: number;
  name: string;
  dataSize: string;
  validity: number;
  price: number;
  operatorId: number;
  /** @nullable */
  operatorName?: string | null;
  /** @nullable */
  operatorSlug?: string | null;
  /** @nullable */
  operatorColor?: string | null;
  active: boolean;
  createdAt: string;
}

export interface BundleInput {
  name: string;
  dataSize: string;
  validity: number;
  price: number;
  operatorId: number;
  active?: boolean;
}

export interface BundleUpdate {
  name?: string;
  dataSize?: string;
  validity?: number;
  price?: number;
  active?: boolean;
}

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  mtn_momo: "mtn_momo",
  orange_money: "orange_money",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  bundleId?: number;
  phoneNumber: string;
  paymentMethod: OrderPaymentMethod;
  status: OrderStatus;
  totalAmount: number;
  /** @nullable */
  transactionId?: string | null;
  bundle?: Bundle;
  createdAt: string;
}

export type OrderInputPaymentMethod =
  (typeof OrderInputPaymentMethod)[keyof typeof OrderInputPaymentMethod];

export const OrderInputPaymentMethod = {
  mtn_momo: "mtn_momo",
  orange_money: "orange_money",
} as const;

export interface OrderInput {
  bundleId: number;
  phoneNumber: string;
  paymentMethod: OrderInputPaymentMethod;
}

export type PaymentInputPaymentMethod =
  (typeof PaymentInputPaymentMethod)[keyof typeof PaymentInputPaymentMethod];

export const PaymentInputPaymentMethod = {
  mtn_momo: "mtn_momo",
  orange_money: "orange_money",
} as const;

export interface PaymentInput {
  paymentMethod: PaymentInputPaymentMethod;
}

export interface PaymentResult {
  success: boolean;
  transactionId: string;
  /** @nullable */
  message?: string | null;
  order: Order;
}

export type RevenueStatsRevenueByOperatorItem = {
  operatorName: string;
  revenue: number;
  orders: number;
};

export interface RevenueStats {
  totalRevenue: number;
  totalOrders: number;
  paidOrders: number;
  revenueByOperator: RevenueStatsRevenueByOperatorItem[];
}

export interface OrderStats {
  total: number;
  pending: number;
  paid: number;
  failed: number;
  cancelled: number;
  recentOrders: Order[];
}

export interface PopularBundle {
  bundle: Bundle;
  totalOrders: number;
  totalRevenue: number;
}

export type ListBundlesParams = {
  operatorId?: number;
  active?: boolean;
};