# lib/api-zod/src/generated/types/bundleUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Good Deal - Telecom Bundle Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface BundleUpdate {
  name?: string;
  dataSize?: string;
  validity?: number;
  price?: number;
  active?: boolean;
}