# lib/api-zod/src/generated/types/operator.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Good Deal - Telecom Bundle Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Operator {
  id: number;
  name: string;
  slug: string;
  color: string;
  /** @nullable */
  logoUrl: string | null;
  active: boolean;
}