# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Good Deal - Telecom Bundle Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Bundle } from "./bundle";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  bundleId?: number;
  phoneNumber: string;
  paymentMethod: OrderPaymentMethod;
  status: OrderStatus;
  totalAmount: number;
  /** @nullable */
  transactionId?: string | null;
  bundle?: Bundle;
  createdAt: Date;
}