# lib/api-zod/src/generated/types/popularBundle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Good Deal - Telecom Bundle Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Bundle } from "./bundle";

export interface PopularBundle {
  bundle: Bundle;
  totalOrders: number;
  totalRevenue: number;
}